# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend server for Vizen Application.",
        "title": "Vizen API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte Vizen",
            "email": "support@vizen.app"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/invites/validate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates a visitor's QR code token. Checks if the token exists, belongs to the condominium, is within the valid time window, and has not been revoked. Records the entry in access logs upon success.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invites"
                ],
                "summary": "Validate Access Token (QR Code)",
                "parameters": [
                    {
                        "description": "Token Validation Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_controllers.ValidateInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_controllers.ValidateInviteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Payload",
                        "schema": {
                            "$ref": "#/definitions/github_com_Bellorico323_vizen_internal_api_common.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_Bellorico323_vizen_internal_api_common.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Business Rule Violation (Expired, Revoked, Not Started) or Permission Denied",
                        "schema": {
                            "$ref": "#/definitions/github_com_Bellorico323_vizen_internal_api_common.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Invite/Token not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Bellorico323_vizen_internal_api_common.ErrResponse"
                        }
                    },
                    "422": {
                        "description": "Validation Failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_Bellorico323_vizen_internal_api_common.ValidationErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Bellorico323_vizen_internal_api_common.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api_controllers.ValidateInviteRequest": {
            "type": "object",
            "required": [
                "condominiumId",
                "token"
            ],
            "properties": {
                "condominiumId": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "api_controllers.ValidateInviteResponse": {
            "type": "object",
            "properties": {
                "invite": {
                    "$ref": "#/definitions/github_com_Bellorico323_vizen_internal_store_pgstore.GetInviteByTokenRow"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Bellorico323_vizen_internal_api_common.ErrResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Bellorico323_vizen_internal_api_common.ValidationErrResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validator.ErrorResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_Bellorico323_vizen_internal_store_pgstore.GetInviteByTokenRow": {
            "type": "object",
            "properties": {
                "apartment_id": {
                    "type": "string"
                },
                "apartment_number": {
                    "type": "string"
                },
                "block": {
                    "type": "string"
                },
                "condominium_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "guest_name": {
                    "type": "string"
                },
                "guest_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "issued_by": {
                    "type": "string"
                },
                "resident_name": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "validator.ErrorResponse": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}